Regression test for importing from io
Works with sbt-ensime 1.12.15 (or no ensime plugin), but with
sbt-ensime 2.0.0 and 2.0.1 this fails with
/build.sbt:1: error: object gatling is not a member of package sbt.io
import io.gatling.sbt.GatlingKeys._

A workaround is to use an absolute import, so that io doesn't resolve
to sbt.io.

Ideally, sbt-ensime shouldn't import sbt.io into the scope of the
build file, as that should be a per-project decision, whereas
the choice to use ensime should be per-user.